src/requirements.txt
setuptools==82.0.0
uv==0.10.4
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
scikit-learn==1.4.2
pillow==10.3.0deactivate